@@ -5,8 +5,12 @@
   "devDependencies": {
     "enzyme": "3.2.0",
     "enzyme-adapter-react-16": "^1.1.0",
+    "eslint": "^6.6.0",
+    "eslint-config-prettier": "^6.5.0",
+    "eslint-plugin-react": "^7.16.0",
     "faker": "^4.1.0",
-    "react-scripts": "^3.0.1",
+    "prettier": "^1.18.2",
+    "react-scripts": "3.2.0",
     "react-test-renderer": "16.0.0",
     "redux-devtools": "^3.5.0",
     "redux-devtools-dock-monitor": "^1.1.3",
@@ -20,22 +24,22 @@
     "bootstrap": "^3.3.7",
     "chain-function": "^1.0.1",
     "cheerio": "^0.22.0",
-    "dialog-polyfill": "^0.4.10",
+    "dialog-polyfill": "^0.5.0",
     "firebase": "5.5.6",
-    "history": "4.7.2",
+    "history": "4.10.1",
     "idb-keyval": "^3.2.0",
     "lodash": "^4.17.4",
     "lodash.clonedeep": "^4.5.0",
     "moment-timezone": "^0.5.25",
     "prop-types": "15.6.0",
-    "react": "16.8.1",
+    "react": "16.12.0",
     "react-bootstrap": "0.32.0",
-    "react-dom": "16.8.1",
+    "react-dom": "16.12.0",
     "react-notification-system": "0.2.16",
     "react-portal": "^4.0.0",
-    "react-redux": "^6.0.0",
+    "react-redux": "7.1.3",
     "react-router-bootstrap": "^0.24.2",
-    "react-router-dom": "4.2.2",
+    "react-router-dom": "5.1.2",
     "react-scrollchor": "5.0.2",
     "react-sizes": "1.0.0",
     "react-stockcharts": "^0.7.8",
