@@ -2,8 +2,8 @@
 
 import { putLast20Closes } from './chartDataCache'
 
-const buildLast20Closes = (symbol, daylyPriceData) => {
-  let last20Prices = daylyPriceData.slice(-20)
+const buildLast20Closes = (symbol, dailyPriceData) => {
+  let last20Prices = dailyPriceData.slice(-20)
   let last20Closes = last20Prices.map((obj) => {
     return { close: obj.close, date: obj.date }
   })
